haredoc: show functions with same siguature as alias
When viewing a single type declaration, haredoc checked for other
declarations whose type was the alias being viewed, as well as functions
returning said type. However, it didn't check for function declarations
with the same signature as a function type being aliased. This commit
changes that.

For example: try viewing docs for hare::unparse::synfunc or
rt::abort_handler. With this patch applied, the former will also show
syn_nowrap and syn_wrap, and the latter will also show _abort.

The logic isn't perfect, but I think this is the best it can get without
a type checker (unless significant complexity is added, and even then it
still wouldn't cover everything). In practice, this should be good
enough.

Signed-off-by: Sebastian <sebastian@sebsite.pw>