time::date: fix realize(); rm lack error type
It is somewhat apparent that one would like a clear error message when
realize() returns [[insufficient]]. What precisely is missing from a
given virtual date? Since there are multiple possible valid answers,
this is not a trivial question.

As implemented, [[lack]] was not a good solution to this problem.
We close on this issue for another day.

Signed-off-by: Byron Torres <b@torresjrjr.com>