bufio: add managed bufio::stream modes
This allows the caller to tune the degree to which bufio::stream has
ownership over its resources and will clean them up when the stream is
closed (i.e. freeing the read/write buffers and/or closing the
underlying file).

os::stdout et al are also updated to configure bufio to close the
underlying file when closed, so io::close(os::stdout) will do what the
user expects.

Signed-off-by: Drew DeVault <drew@ddevault.org>