encoding::utf8, strings: add position function
Sometimes, it's desirable to obtain the current position of an iterator.
Prior to this commit, the only ""correct"" way to do this was to create
an iterator for the start of the string, and take the length of the
result of strings::slice. It's much easier to just directly access
iter.dec.offs, and some code did exactly this, even though that field is
internal.

This commit introduces the strings::position function, making this more
ergonomic. encoding::utf8::position has also been added for parity. The
docs note that other functions like iterstr and slice are often
preferable.

Signed-off-by: Sebastian <sebastian@sebsite.pw>