path: clarify MAX docs
"Maximum length of a file path for this platform" is oversimplified and
arguably incorrect. The new docs better explain the actual meaning of
this constant.

The initializer is also changed to be `rt::PATH_MAX - 1` on all
platforms. The value itself is unchanged, but the constant is no longer
hard-coded.

Signed-off-by: Sebastian <sebastian@sebsite.pw>
Signed-off-by: Drew DeVault <drew@ddevault.org>