crypto::ed25519: new verification api
Common signature algorithms, like the others in crypto::, require a
message hash, that will be signed or verified. ed25519 differs in this
aspect, since it needs the whole message as input. Hashing the message
has the advantage to not beeing forced to have the whole message in
memory.

This commit introduces changes the verify method to verify_buf and
introduces the verifier function, that creates a hash the message can be
written into and verified with verify. This allows verification of
message read from an io::stream. The old variant is available through
the verify_buf helper function.

* crypto::ed25519::verify: rename to verify_buf

Breaking-Change: crypto::ed25519
Signed-off-by: Armin Preiml <apreiml@strohwolke.at>