Makefile: build configured hare(1) from all target
Previously when following the recommended boostrapping steps of:
`$ make` then `# make install`, users with:

1. git version >=2.35.2
2. the hare/ repo owned by a regular user
and
3. the hare/ repo not configured as a `safe.directory`

might not have meaningful version strings embedded in the installed
hare(1) binary. This was because the $BINOUT/hare-install target would
only be built by root, causing the git(1) commands in scripts/version to
fail with "fatal: detected dubious ownership in repository" or similar.

Signed-off-by: Rosie K Languet <rkl@rosiesworkshop.net>